+++ Panther/src/AST/Parser.cpp
@@ -3202,9 +3202,17 @@ namespace pcit::panther{
 					}else{
 						output = this->source.ast_buffer.createPostfix(output.value(), this->reader.next());
 					}
-
 				} break;
 
+				case Token::lookupKind("..."): {
+					if constexpr(TERM_KIND == TermKind::EXPR){
+						output = this->source.ast_buffer.createPostfix(output.value(), this->reader.next());
+						
+					}else{
+						should_continue = false;
+						break;
+					}
+				} break;
 
 				case Token::lookupKind("<{"): {
 					if(this->assert_token(Token::lookupKind("<{")).isError()){ return Result::Code::ERROR; }
@@ -3291,11 +3299,7 @@ namespace pcit::panther{
 				} break;
 
 				case Token::lookupKind("["): {
-					if constexpr(TERM_KIND == TermKind::EXPLICIT_TYPE || TERM_KIND == TermKind::AS_TYPE){
-						should_continue = false;
-						break;
-
-					}else{
+					if constexpr(TERM_KIND == TermKind::EXPR){
 						const Token::ID open_bracket = this->reader.next();
 
 						auto indices = evo::SmallVector<AST::Node>();
@@ -3330,6 +3334,10 @@ namespace pcit::panther{
 
 						output =
 							this->source.ast_buffer.createIndexer(output.value(), std::move(indices), open_bracket);
+
+					}else{
+						should_continue = false;
+						break;
 					}
 				} break;